distributed_tensorflow/requirements.txt
keystoneauth1==3.10.0
python-heatclient==1.16.1
